requirements.txt
alembic==1.12.0
amqp==5.2.0
aniso8601==9.0.1
async-timeout==4.0.3
bidict==0.22.1
billiard==4.2.0
blinker==1.6.3
bunch @ git+https://github.com/yigitbasalma/bunch.git@aa767ad44e86b0ab77af9f566f6fda392f23c516
celery==5.3.5
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cryptography==41.0.5
dnspython==2.4.2
docopt==0.6.2
email-validator==2.0.0.post2
Flask==3.0.0
Flask-Admin==1.6.1
Flask-Authorize==0.2.6
Flask-Login==0.6.2
Flask-Migrate==4.0.5
Flask-RESTful==0.3.10
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
h11==0.14.0
idna==3.4
importlib-metadata==6.8.0
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.3
Mako==1.2.4
MarkupSafe==2.1.3
numpy==1.26.1
pandas==2.1.3
pipreqs==0.4.13
prompt-toolkit==3.0.40
pycparser==2.21
PyMySQL==1.1.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-engineio==4.8.0
python-socketio==5.10.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
simple-websocket==1.0.0
six==1.16.0
slack-sdk==3.23.1
SQLAlchemy==2.0.22
timeago==1.0.16
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
vine==5.1.0
wcwidth==0.2.9
wsproto==1.2.0
WTForms==3.1.0
yarg==0.1.9
zipp==3.17.0
pymsteams==0.2.2
